requirements.txt

faiss-cpu==1.13.2
numpy==1.26.4
requests==2.32.1
pypdf==3.17.0       
easyocr==1.7.2
groq==0.11.0
opencv-python-headless==4.8.1.78
jina==3.33.0
altair==5.5.0
blinker==1.9.0
cachetools==5.5.2
click==8.3.1
importlib-metadata==6.11.0
packaging==23.2
pandas==2.3.3
streamlit>=1.28.0,<2.15
Pillow>=12.1
